app/requirements.txt
pandas
megadetector
Flask
minio
yolov5
python-dotenv